Issue 37: port php project (#38)
* Issue_2: newsService, dtos, controller, validator, repo

* issue_21: make embeddable entity, try to reduce code duplicates

* issue_21: fix

* issue_4:conflict feature

* issue_4: PR fixes

* issue_23: client version feature

* issue_24:refactor packages

* issue_25: user feature

* issue_26: localities layers

* issue_31: comments feature + moderation endpoint

* issue_31: small type fix

* issue_33: moderation dashboard endpoint

* issue_32: references endpoints

* pr_fixes: huge request dto refactor

* pr_fixes: huge validation refactor

* iss39: remove POST list endpoints

* camelCase requests

* bugfix: news validation

* bugfix: favourites filter

* camel case parameters

* fix: stateless session

* pr_fixes: remove static imports

* remove dto validator tests

* pr_fixes: remove comments feature

* pr_fixes: update readme

* pr_fixes: consider static methods instead of contructors

* pr_fixes: remove comment from entityGraph

* bugfix: news repo

* pr_fixes: huge refactor

+++ src/main/java/ru/smartel/strike/entity/Photo.java
@@ -3,7 +3,12 @@ package ru.smartel.strike.entity;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import java.time.LocalDateTime;
 
 @Entity
@@ -13,7 +18,7 @@ public class Photo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
-    private int id;
+    private long id;
 
     @CreationTimestamp
     @Column(name = "created_at")
@@ -33,7 +38,7 @@ public class Photo {
         this.url = url;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
